# chore(repo): ESLint refactor + speed up 70%