--- a/packages/coinjoin/src/backend/CoinjoinMempoolController.ts
+++ b/packages/coinjoin/src/backend/CoinjoinMempoolController.ts
@@ -143,9 +143,7 @@ export class CoinjoinMempoolController {
                 .concat(change)
                 .map(({ address }) => getAddressScript(address, this.network));
 
-            // eslint-disable-next-line no-await-in-loop
             await promiseAllSequence(
-                // eslint-disable-next-line no-loop-func
                 filters.map(([txid, matchAny], index) => async () => {
                     if (matchAny(scripts)) await addTx(txid);
                     if (progressCooldown())