--- a/packages/coinjoin/src/backend/scanAccount.ts
+++ b/packages/coinjoin/src/backend/scanAccount.ts
@@ -43,7 +43,7 @@ export const scanAccount = async (
     const txs = new Set<BlockbookTransaction>();
 
     const everyFilter = filters.getFilterIterator({ checkpoints }, { abortSignal, onProgressInfo });
-    // eslint-disable-next-line no-restricted-syntax
+
     for await (const { blockHash, blockHeight, filter, filterParams } of everyFilter) {
         const isMatch = getMultiFilter(filter, filterParams);
         const scripts = addresses.receive.concat(addresses.change).map(({ script }) => script);