--- a/packages/coinjoin/src/global.d.ts
+++ b/packages/coinjoin/src/global.d.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable max-classes-per-file */
 
 declare module 'golomb' {
     class Golomb {