--- a/packages/components/src/components/assets/CoinLogo/coins.ts
+++ b/packages/components/src/components/assets/CoinLogo/coins.ts
@@ -1,4 +1,3 @@
-/* eslint-disable global-require */
 export const COINS = {
     ada: require('../../../images/coins/ada.svg'),
     bch: require('../../../images/coins/bch.svg'),