--- a/packages/components/src/components/assets/TrezorLogo/trezorLogos.ts
+++ b/packages/components/src/components/assets/TrezorLogo/trezorLogos.ts
@@ -1,4 +1,3 @@
-/* eslint-disable global-require */
 export const LOGOS: { [key: string]: any } = {
     HORIZONTAL: require('../../../images/trezor_logo_horizontal.svg'),
     VERTICAL: require('../../../images/trezor_logo_vertical.svg'),