--- a/packages/connect-explorer/src/components/Header.tsx
+++ b/packages/connect-explorer/src/components/Header.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable jsx-a11y/control-has-associated-label */
-
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 import Devices from './Devices';