--- a/packages/connect-explorer/src/components/PageAbout.tsx
+++ b/packages/connect-explorer/src/components/PageAbout.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/no-danger */
 import { useDocs } from '../hooks/useDocs';
 
 export const About = () => {