--- a/packages/connect-explorer/src/components/PageChangelog.tsx
+++ b/packages/connect-explorer/src/components/PageChangelog.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/no-danger */
 import { useDocs } from '../hooks/useDocs';
 
 export const Changelog = () => {