--- a/packages/connect-explorer/src/data/methods/bitcoin/signTransaction.paytoaddress.ts
+++ b/packages/connect-explorer/src/data/methods/bitcoin/signTransaction.paytoaddress.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-bitwise */
 import { select } from './common';
 
 const name = 'signTransaction';