--- a/packages/connect-explorer/webpack/prod.webpack.config.ts
+++ b/packages/connect-explorer/webpack/prod.webpack.config.ts
@@ -80,7 +80,6 @@ const config: webpack.Configuration = {
             minify: false,
         }),
         new webpack.DefinePlugin({
-            // eslint-disable-next-line no-underscore-dangle
             'process.env.__TREZOR_CONNECT_SRC': JSON.stringify(process.env.__TREZOR_CONNECT_SRC),
             'process.env.COMMIT_HASH': JSON.stringify(commitHash),
             'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),