--- a/packages/connect-explorer/webpack/webextension.webpack.config.ts
+++ b/packages/connect-explorer/webpack/webextension.webpack.config.ts
@@ -166,7 +166,6 @@ const config: webpack.Configuration = {
             ],
         }),
         new webpack.DefinePlugin({
-            // eslint-disable-next-line no-underscore-dangle
             'process.env.__TREZOR_CONNECT_SRC': JSON.stringify(process.env.__TREZOR_CONNECT_SRC),
             'process.env.COMMIT_HASH': JSON.stringify(commitHash),
         }),