--- a/packages/connect-webextension/src/channels/serviceworker-window.ts
+++ b/packages/connect-webextension/src/channels/serviceworker-window.ts
@@ -77,7 +77,7 @@ export class ServiceWorkerWindowChannel<
                 }
 
                 // TODO: not completely sure that is necessary to prevent self origin communication sometimes.
-                // eslint-disable-next-line no-restricted-globals
+
                 if (origin === self.origin && !allowSelfOrigin) {
                     return;
                 }