--- a/packages/connect-webextension/src/popup.ts
+++ b/packages/connect-webextension/src/popup.ts
@@ -110,7 +110,6 @@ export class PopupManager extends EventEmitter {
                             active: true,
                         },
                         tabs => {
-                            // eslint-disable-next-line prefer-destructuring
                             this.popupWindow = tabs[0];
                             this.injectContentScript(this.popupWindow.id!);
                         },