--- a/packages/connect/e2e/common.setup.js
+++ b/packages/connect/e2e/common.setup.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-named-as-default
 import TrezorConnect from '../src';
 import { versionUtils } from '@trezor/utils';
 import { UI } from '../src/events';