--- a/packages/connect/src/api/getAccountInfo.ts
+++ b/packages/connect/src/api/getAccountInfo.ts
@@ -264,7 +264,7 @@ export default class GetAccountInfo extends AbstractMethod<'getAccountInfo', Req
                     if (this.hasBundle) {
                         responses.push(null);
                         sendProgress(i, null, error.message);
-                        // eslint-disable-next-line no-continue
+
                         continue;
                     } else {
                         throw error;
@@ -326,7 +326,7 @@ export default class GetAccountInfo extends AbstractMethod<'getAccountInfo', Req
                 if (this.hasBundle) {
                     responses.push(null);
                     sendProgress(i, null, error.message);
-                    // eslint-disable-next-line no-continue
+
                     continue;
                 } else {
                     throw error;