--- a/packages/connect/src/data/analyticsInfo.ts
+++ b/packages/connect/src/data/analyticsInfo.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-case-declarations */
 // import { EventType } from '@trezor/connect-analytics';
 import { CoreEventMessage, UI_REQUEST } from '../events';
 import type { Device } from '../types';