--- a/packages/e2e-utils/src/mocks/dropbox.ts
+++ b/packages/e2e-utils/src/mocks/dropbox.ts
@@ -1,4 +1,4 @@
-/* eslint-disable camelcase, no-console */
+/* eslint-disable no-console */
 
 import express, { Express } from 'express';