--- a/packages/ipc-proxy/src/__tests__/mockedElectron.ts
+++ b/packages/ipc-proxy/src/__tests__/mockedElectron.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 import { EventEmitter } from 'events';
 
 class IpcRendererMock extends EventEmitter {