--- a/packages/suite-desktop-api/src/validation.ts
+++ b/packages/suite-desktop-api/src/validation.ts
@@ -7,7 +7,7 @@ type OptionalPrimitive = Primitive | [Primitive, boolean];
 export const isPrimitive = (type: OptionalPrimitive, value: any) => {
     const [t, optional] = Array.isArray(type) ? type : [type];
     if (value == null && optional) return true;
-    // eslint-disable-next-line valid-typeof
+
     return typeof value === t;
 };