… .github/workflows/validation.yml
… package.json
… packages/blockchain-link/tests/integration/worker/index.ts
… packages/blockchain-link/tests/unit/workers.test.ts
… packages/coinjoin/tests/backend/CoinjoinBackendClient.test.ts
… packages/coinjoin/tests/backend/CoinjoinFilterController.test.ts
… packages/coinjoin/tests/tools/anonymity-test.ts
… packages/coinjoin/tests/tools/benchmark-test.ts
… packages/connect-plugin-ethereum/__tests__/index.test.ts
… packages/connect-popup/e2e/tests/browser-support.test.ts
… packages/connect-popup/e2e/tests/methods.test.ts
… packages/connect/e2e/tests/api/init.test.ts
… packages/connect/e2e/tests/device/authenticateDevice.test.ts
… packages/connect/e2e/tests/device/authorizeCoinjoin.test.ts
… packages/connect/e2e/tests/device/cancelCoinjoinAuthorization.test.ts
… packages/connect/e2e/tests/device/checkFirmwareAuthenticity.test.ts
… packages/connect/e2e/tests/device/methods.test.ts
… packages/connect/e2e/tests/device/override.test.ts
… packages/connect/e2e/tests/device/passphrase.test.ts
… packages/connect/e2e/tests/device/setBusy.test.ts
… packages/connect/e2e/tests/device/unlockPath.test.ts
… packages/suite-desktop-core/e2e/tests/coinjoin.test.ts
… packages/suite-storage/src/web/index.ts
… packages/suite-web/e2e/run_tests.ts
… packages/suite/src/actions/suite/__tests__/initAction.test.ts
… packages/suite/src/actions/suite/__tests__/routerActions.test.ts
… packages/suite/src/actions/suite/metadataProviderActions.ts
… packages/suite/src/actions/wallet/moveLabelsForRbfActions.ts
… packages/suite/src/components/suite/DeviceDisplay.tsx
… packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/TransactionReviewOutputList/TransactionReviewOutputList.tsx